+++ CHANGELOG.md
@@ -96,6 +96,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * OpenShell3D: to_step_face_ids
 * Contour2D: repair_cut_contour
 * Circle2D: cut_by_line
+* Contour3D: linesegment_intersections, line_intersections
+* Circle3D: primitives: [Arc3D, Arc3D], get_primitives, abscissa, linesegment_intersections
+* Arc3D: to_circle3D, linesegment_intersections
+* 
 
 ### Fixed
 
@@ -116,6 +120,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Surface2D: cut_by_line
 * Circle3D: to_step
 * ArcEllipse3D.to_2d()
+* Circle3D: point_belongs
+* Circle3D: discretization_points
 
 
 ### Performance improvements